requirements.txt
celery==4.2.1
Django==2.1.5
django-celery-beat==1.4.0
django-celery-results==1.0.4
djangorestframework==3.9.0
djoser==1.3.2
redis==3.0.1
requests==2.21.0
